Test/test_main.py
# Create test_main.py with detailed instructions and test cases

test_main_content = """
\"\"\"
Test Script: test_main.py
Description: Tests the orchestration and integration logic in `main.py`.

Run Instructions:
    pytest test_main.py

Preparation Instructions:
    - Ensure all modules (`main.py`, `DirectoryScanner`, `FileProcessor`, etc.) are updated and functional.
    - Mock any external dependencies such as GitHub or database connections to avoid impacting production systems.

Reset Instructions:
    - Remove any test artifacts (e.g., test files or directories) created during the test.
\"\"\"

import asyncio
from pathlib import Path
from unittest.mock import patch, AsyncMock

@patch("main.ensure_repository_synced")
@patch("main.DirectoryScanner.scan", return_value=[Path("test_file.md")])
@patch("main.FileProcessor.validate_and_process_file", return_value={"success": True})
def test_main_workflow(mock_process, mock_scan, mock_sync):
    from main import process_files, Config

    # Create test configuration
    config = Config.load_default()

    # Run the main process function
    asyncio.run(process_files(config))

    # Verify repository sync was called
    mock_sync.assert_called_once()

    # Verify scanning was performed
    mock_scan.assert_called_once()

    # Verify file processing was called
    mock_process.assert_called_once_with(Path("test_file.md"))

def test_main_error_handling():
    from main import process_files, Config

    # Simulate failure in processing
    with patch("main.FileProcessor.validate_and_process_file", return_value={"success": False, "error": "Test Error"}):
        config = Config.load_default()

        # Run the main process and expect it to handle errors gracefully
        try:
            asyncio.run(process_files(config))
        except Exception as e:
            assert False, f"Error handling failed: {e}"
"""

# Save the test_main.py script to a file for download
test_main_path = "/mnt/data/test_main.py"
with open(test_main_path, "w") as file:
    file.write(test_main_content)

test_main_path
